refactor: move window manamgent related files (#1524)
We should not have files directly under electron folder but we do.

In this PR I move all window management-related files to their own folder and also correct the test files path.
