+++ src/electron/window-management/window-frame-listener.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import { BrowserWindow } from 'electron';
-import { WindowStateActionCreator } from './flux/action-creator/window-state-action-creator';
+import { WindowStateActionCreator } from '../flux/action-creator/window-state-action-creator';
 
 export class WindowFrameListener {
     constructor(private readonly windowStateActionsCreator: WindowStateActionCreator, private readonly browserWindow: BrowserWindow) {}
