+++ src/electron/window-management/window-frame-updater.ts
